+++ src/library/DIPS.Mobile.UI/API/Builder/Android/AppHostBuilderExtensions.cs
@@ -1,8 +1,8 @@
 using DIPS.Mobile.UI.API.Library;
-using DIPS.Mobile.UI.Components.Shell.Android;
 using DIPS.Mobile.UI.Components.Toolbars.Android;
 using Microsoft.Maui.LifecycleEvents;
-using Shell = DIPS.Mobile.UI.Components.Shell.Shell;
+using ContentPage = DIPS.Mobile.UI.Components.Pages.ContentPage;
+using ContentPageHandler = DIPS.Mobile.UI.Components.Pages.Android.ContentPageHandler;
 
 namespace DIPS.Mobile.UI.API.Builder;
 
@@ -12,8 +12,7 @@ public static partial class AppHostBuilderExtensions
     {
         handlers.AddHandler(typeof(DIPS.Mobile.UI.Components.Searching.Android.IndeterminateProgressBar), typeof(DIPS.Mobile.UI.Components.Searching.Android.IndeterminateProgressBarHandler));
         handlers.AddHandler<DUIToolbar, DUIToolbarHandler>();
-        handlers.AddHandler<Shell, ShellRenderer>();
-        handlers.AddHandler<Toolbar, ToolbarHandler>();
+        handlers.AddHandler<ContentPage, ContentPageHandler>();
     }
 
     static partial void ConfigurePlatformLifecycleEvents(ILifecycleBuilder events)
