--- src/library/DIPS.Mobile.UI/Components/Pages/dotnet/ContentPageHandler.cs
@@ -1,7 +1,0 @@
-using Microsoft.Maui.Handlers;
-
-namespace DIPS.Mobile.UI.Components.Pages;
-
-public partial class ContentPageHandler : PageHandler
-{
-}
