+++ src/library/DIPS.Mobile.UI/Components/Shell/ShellRenderer.cs
@@ -1,0 +1,6 @@
+namespace DIPS.Mobile.UI.Components.Shell;
+
+public partial class ShellRenderer
+{
+    
+}
