+++ src/library/DIPS.Mobile.UI/Platforms/iOS/CustomShellHandler.cs
@@ -1,0 +1,6 @@
+namespace DIPS.Mobile.UI.Platforms.iOS;
+
+public class CustomShellHandler
+{
+    
+}
